refactor: migrate from oas3-tools to @oas-tools/core
We want to use experimental libraries @digitalbazaar/ed25519-signature-2020 @digitalbazaar/vc, which are written in JS ESM modules. oas3-tools is common.js, and while these two can work together, it is more managable to have all in ESM. oas-tools was available as alternative, so migrate to that.

There was 1 caveat: it resolves the openapi schema as draft04. Using this json schema validator, 'id' and '' are special keywords. Since our schema's contain 'id's, we renamed them to '_id' in the schema. A fix or workaround is still being looked at, as we now not fully conform to the VC standard